Folders restructurized
