+++ frontend/src/components/VisualizerStage/Canvas/VisualizerLine.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Arrow } from 'react-konva';
 
-import { LineProps } from '../interfaces';
+import { LineProps } from '../../interfaces/interfaces';
 
 const Line = ({ key, points }: LineProps): JSX.Element => {
     return <Arrow key={key} strokeWidth={1} stroke="black" fill="black" points={[...points]} />;
