+++ frontend/src/store/features/proof/proofSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { RootState } from '../../store';
 import { processDot, piNodeChildren, piNodeParents, descendants } from './auxi';
 
 export interface NodeInterface {
