+++ frontend/src/store/features/theme/themeSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { RootState } from '../../app/store';
+import { RootState } from '../../store';
 
 export interface ThemeState {
     value: boolean;
